+++ Sources/KeyboardKit/_Deprecated/LocaleKey.swift
@@ -1,0 +1,2 @@
+@available(*, deprecated, renamed: "KeyboardLocale")
+public typealias LocaleKey = KeyboardLocale